refactor: made changes in YouTubeDownloaderView
